Move Bridge model to bridge file
